Minor name changes and removing extra func
