--- client/src/components/FeedbackList/index.js
@@ -1,169 +1,0 @@
-import React, { useState, useEffect } from "react";
-import "./styles.css";
-import modules from "../../db/modules.json";
-import PropTypes from "prop-types";
-import Button from "../Button/Button";
-import timeDifference from "../../helperFunctions/timeDifference";
-
-const FeedbackList = ({ feedbackToShow, student, updateFeedback }) => {
-
-	const [editedFeedback, setEditFeedback] = useState({});
-	const [notEditable, setNotEditable] = useState(true);
-
-	useEffect(() => {
-		setEditFeedback(feedbackToShow);
-	}, [feedbackToShow]);
-
-	//POST updated feedback to DB
-	//When clicked SAVE; UPDATE the data
-	const updateData = (id) => {
-		fetch(`/api/students/${student._id}/feedback/${id}`, {
-			method: "PUT",
-			headers: { "Content-type": "application/json" },
-			body: JSON.stringify(editedFeedback) })
-			.then( (res) => res.json())
-			.catch((error) => console.log(error));
-	};
-
-	//When clicked SAVE, CHECK if field is empty
-	const saveFeedback = (id) => {
-		if (!editedFeedback.title) {
-			return alert("Please add a title");
-		} else if (!editedFeedback.text) {
-			return alert("Please add your feedback");
-		} else if (!editedFeedback.mentor) {
-			return alert("Please add your name");
-		}
-		updateData(id);
-	};
-
-	//DELETE selected feedback and update data
-	const handleDelete = (feedbackId) => {
-		fetch(`/api/students/${student._id}/${feedbackId}`, {
-			method: "DELETE",
-			headers: { "Content-type": "application/json" },
-		})
-			.then((res) => res.json())
-			.catch((error) => console.log(error));
-	};
-
-	// Populate new object with edited data editable
-	const handleEdit = (e) => {
-		setEditFeedback({ ...editedFeedback, [e.target.name]: e.target.value });
-	};
-
-	return (
-		<div className="previous-feedback-list">
-			{Object.keys(editedFeedback).length
-			&& Object.keys(editedFeedback).map((property, index) => {
-				if (property === "module" && notEditable ) {
-					return (
-						<div className="previous-feedback-module" key={index}>
-							<p className="feedback-input-head">Feedback module: </p>
-							<p>{editedFeedback[property]}</p>
-						</div> );
-				}
-				if (property === "module" && !notEditable ) {
-					return (
-						<div className="previous-feedback-module" key={index}>
-							<p className="feedback-input-head">Feedback module: </p>
-							<select
-								name="module"
-								disabled={notEditable}
-								value={editedFeedback[property]}
-								onChange={handleEdit}>
-								{modules.map((module,index) =>
-									<option key={index} value={module.name}> {module.name}</option>
-								)}
-							</select>
-						</div>
-					);
-				}
-				if (property === "title") {
-					return (
-						<div className="previous-feedback-title" key={index}>
-							<p className="feedback-input-head">Feedback title</p>
-							<input
-								className="feedback-title-input"
-								onChange={handleEdit}
-								name="title"
-								value={editedFeedback[property]}
-								disabled={notEditable}
-							/>
-						</div> );
-				}
-				if (property === "text") {
-					return (
-						<div className="previous-feedback-text" key={index}>
-							<p className="feedback-input-head">Feedback</p>
-							<textarea
-								className="previous-feedback-text-input"
-								onChange={handleEdit}
-								name="text"
-								value={editedFeedback[property]}
-								disabled={notEditable}
-							/>
-						</div> );
-				}
-				if (property === "time") {
-					return (
-						<div className="previous-feedback-time" key={index}>
-							<input
-								className="previous-feedback-time-input"
-								onChange={handleEdit}
-								name="time"
-								value={timeDifference(Date.now(), editedFeedback[property])}
-								disabled
-							/>
-						</div>
-					);
-				}
-				if (property === "mentor") {
-					return (
-						<div className="previous-feedback-mentor" key={index}>
-							<p className="feedback-input-head">Given by: </p>
-							<input
-								className="previous-feedback-mentor-input"
-								onChange={handleEdit}
-								name={property}
-								value={editedFeedback[property]}
-								disabled
-							/>
-						</div>
-					);
-				}
-			})}
-			<div className="edit-delete-buttons">
-				<Button
-					content={notEditable ? "Edit" : "Save"}
-					handleClick={() => {
-						if (notEditable) {
-							setNotEditable(false);
-						} else {
-							saveFeedback(editedFeedback.id);
-							setNotEditable(true);
-						}
-					}}
-				/>
-				<div>
-					<Button
-						content="Delete"
-						handleClick={() => {
-							alert("DELETE");
-							handleDelete(editedFeedback.id);
-							updateFeedback();
-						}}
-					/>
-				</div>
-			</div>
-		</div>
-	);
-};
-
-FeedbackList.propTypes = {
-	student: PropTypes.object.isRequired,
-	feedbackToShow: PropTypes.object.isRequired,
-	updateFeedback: PropTypes.func.isRequired,
-};
-
-export default FeedbackList;
